refactor(core): Make native Python runner compatible with launcher
