+++ packages/@n8n/task-runner-python/src/message_types/runner.py
@@ -1,7 +1,7 @@
 from dataclasses import dataclass
 from typing import List, Literal, Union, Any, Dict
 
-from ..constants import (
+from src.constants import (
     RUNNER_INFO,
     RUNNER_TASK_ACCEPTED,
     RUNNER_TASK_DONE,
